move more into variable-plugin
